[MAPS3D-952] use uint16 for secondary icon texture coordinates (internal-620)
* [MAPS3D-952] use uint16 for secondary icon texture coordinates
* update dynamic style
* fix 3d playground debug page UI controllers with latest 3d style (HSV lights)

+++ src/data/bucket/symbol_attributes.js
@@ -23,7 +23,7 @@ export const placementOpacityAttributes: StructArrayLayout = createLayout([
 ], 4);
 
 export const iconTransitioningAttributes: StructArrayLayout = createLayout([
-    {name: 'a_texb', components: 2, type: 'Uint8'}
+    {name: 'a_texb', components: 2, type: 'Uint16'}
 ]);
 
 export const collisionVertexAttributes: StructArrayLayout = createLayout([
